requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
